# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@karimastore.com
    name: Karima Store API Support
  description: |-
    Karima Store E-commerce API with JWT Authentication

    ## Authentication

    This API uses **JWT (JSON Web Token)** for authentication.

    ### For Web/Browser Clients:
    1. Login via /api/v1/auth/login endpoint
    2. Session cookie (jwt) will be set automatically
    3. Make API requests with the cookie included

    ### For API/Mobile Clients:
    1. Obtain JWT token from /api/v1/auth/login endpoint
    2. Include token in requests:
    - Method 1: Authorization: Bearer <jwt_token>
    - Method 2: X-Session-Token: <jwt_token> header

    ### Authorization Levels:
    - **Public**: No authentication required (GET endpoints for browsing)
    - **Authenticated**: Valid JWT token required
    - **Admin**: Valid JWT token + admin role
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Karima Store API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get list of all users (admin only)
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by ID (admin or self)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Get user by ID
      tags:
      - users
  /users/{id}/activate:
    put:
      consumes:
      - application/json
      description: Activate a user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Activate user
      tags:
      - users
  /users/{id}/deactivate:
    put:
      consumes:
      - application/json
      description: Deactivate a user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Deactivate user
      tags:
      - users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Update a user's role (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role update request
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Update user role
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Get current authenticated user
      tags:
      - users
  /users/stats:
    get:
      consumes:
      - application/json
      description: Get user statistics (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Get user statistics
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"